+++ app/Jobs/SearchAndNotifyUser.php
@@ -7,13 +7,23 @@ use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Support\Facades\Log;
 use JobApis\Jobs\Client\Collection;
 use JobApis\Jobs\Client\JobsMulti;
-use JobApis\JobsToMail\Filters\RecruiterJobsFilter;
+use JobApis\JobsToMail\Models\Recruiter;
 use JobApis\JobsToMail\Models\Search;
 use JobApis\JobsToMail\Notifications\JobsCollected;
 
 class SearchAndNotifyUser implements ShouldQueue
 {
     use InteractsWithQueue, Queueable, SerializesModels;
+
+    /**
+     * @var array Jobs collected from providers.
+     */
+    protected $jobs = [];
+
+    /**
+     * @var Recruiter recruiter model
+     */
+    protected $recruiter;
 
     /**
      * @var Search search to conduct
@@ -50,32 +60,33 @@ class SearchAndNotifyUser implements ShouldQueue
      *
      * @return array
      */
-    public function handle(
-        JobsMulti $jobsClient,
-        RecruiterJobsFilter $recruiterFilter
-    ) {
+    public function handle(JobsMulti $jobsClient, Recruiter $recruiter)
+    {
+        // Going to use the recruiter model
+        $this->recruiter = $recruiter;
+
         // Collect jobs based on the Search keyword and location
         $jobsByProvider = $jobsClient->setKeyword($this->search->keyword)
             ->setLocation($this->search->location)
             ->setPage(1, self::MAX_JOBS_FROM_PROVIDER)
             ->getAllJobs();
 
-        // Sort jobs into one array
-        $jobs = $this->getJobsFromCollections($jobsByProvider);
-        $jobs = $this->sortJobs($jobs);
+        // Get jobs in one array
+        $this->jobs = $this->getJobsFromCollections($jobsByProvider);
+
+        // Sort the jobs
+        $this->sortJobs();
 
         // Filter jobs from recruiters
-        $jobs = $recruiterFilter::filter($jobs);
-
-        dd($jobs);
+        $this->filterRecruiterJobs();
 
         // Trigger notification to user
-        if ($jobs) {
-            $this->search->user->notify(new JobsCollected($jobs, $this->search));
+        if ($this->jobs) {
+            $this->search->user->notify(new JobsCollected($this->jobs, $this->search));
         } else {
             Log::info("No jobs found for search {$this->search->id}");
         }
-        return $jobs;
+        return $this->jobs;
     }
 
     /**
@@ -103,6 +114,29 @@ class SearchAndNotifyUser implements ShouldQueue
             }
         );
         return $jobs;
+    }
+
+    /**
+     * Filters out jobs from recruiting companies if the user's prefers it.
+     *
+     * @return void
+     */
+    protected function filterRecruiterJobs()
+    {
+        // Make sure this search wants to filter recruiters
+        if ($this->search->no_recruiters === true) {
+            $this->jobs = array_filter($this->jobs, function ($job) {
+                // Make sure this job has a company
+                if (isset($job->company)) {
+                    // Make sure this company is not a recruiter
+                    $recruiter = $this->recruiter->whereNameLike($job->company)->first();
+                    if ($recruiter) {
+                        return false;
+                    }
+                }
+                return true;
+            });
+        }
     }
 
     /**