+++ app/Models/Recruiter.php
@@ -1,6 +1,7 @@
 <?php namespace JobApis\JobsToMail\Models;
 
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Support\Facades\DB;
 use Ramsey\Uuid\Uuid;
 
 class Recruiter extends Model
@@ -32,5 +33,19 @@ class Recruiter extends Model
         static::creating(function ($model) {
             $model->{$model->getKeyName()} = Uuid::uuid4();
         });
+    }
+
+    /**
+     * Very simple filter to get recruiting firms by name
+     *
+     * @param $query
+     * @param string $name
+     *
+     * @return \Illuminate\Database\Eloquent\Builder
+     */
+    public function scopeWhereNameLike($query, $name = null)
+    {
+        $where = "to_tsvector('english', name) @@ plainto_tsquery('english', ?)";
+        return $query->whereRaw($where, [$name]);
     }
 }