+++ changelog.md
@@ -15,8 +15,9 @@ All Notable changes to `jobs-to-mail` will be documented in this file.
   - View this user's active searches
 - New page to view a user's searches
 - Filtering out recruiter listings:
-  - `recruiters` database table, seeder, model
-  - Boolean value for filtering recruiters from results of a search 
+  - `recruiters` database table, seeder, model.
+  - Boolean value for filtering recruiters from results of a search.
+  - New filter in SearchAndNotifyUser job to remove recruiter listings if preferred.
 
 ### Fixed
 - Plaintext email job listings were showing up as blank.